Copying binaries on Windows
Windows has a weird relationship with symbolic links which boils
down to requiring administrator rights to create them.

This makes sure that if the host system is a windows system, the
binaries will be copied instead of being symlinked. While this
creates a (minor) storage overhead it immensly simplifies the
developer's life.